Split advanced setting options into a new activity with new fragment and clean up all calsses
