+++ app/src/main/AndroidManifest.xml
@@ -52,6 +52,11 @@
         android:theme="@style/QuranToolBar"
         android:configChanges="keyboardHidden|orientation|screenSize"/>
     <activity
+        android:name=".AdvancedQuranPreferenceActivity"
+        android:theme="@style/QuranToolBar"
+        android:configChanges="keyboardHidden|orientation|screenSize"/>
+
+    <activity
         android:name=".SearchActivity"
         android:launchMode="singleTop">
       <intent-filter>
