+++ app/src/main/java/com/quran/labs/androidquran/QuranPreferenceActivity.java
@@ -1,30 +1,18 @@
 package com.quran.labs.androidquran;
 
-import com.quran.labs.androidquran.service.util.PermissionUtil;
-import com.quran.labs.androidquran.ui.QuranActionBarActivity;
-import com.quran.labs.androidquran.ui.fragment.QuranSettingsFragment;
-import com.quran.labs.androidquran.util.AudioManagerUtils;
-import com.quran.labs.androidquran.util.QuranSettings;
-
-import android.Manifest;
 import android.app.Fragment;
 import android.app.FragmentManager;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
 import android.support.v7.app.ActionBar;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
-import android.widget.Toast;
+
+import com.quran.labs.androidquran.ui.QuranActionBarActivity;
+import com.quran.labs.androidquran.ui.fragment.QuranSettingsFragment;
+import com.quran.labs.androidquran.util.AudioManagerUtils;
 
 public class QuranPreferenceActivity extends QuranActionBarActivity {
-
-  private static final String SI_LOCATION_TO_WRITE = "SI_LOCATION_TO_WRITE";
-  private static final int REQUEST_WRITE_TO_SDCARD_PERMISSION = 1;
-
-  private String mLocationToWrite;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -42,9 +30,6 @@ public class QuranPreferenceActivity extends QuranActionBarActivity {
 
     AudioManagerUtils.clearCache();
 
-    if (savedInstanceState != null) {
-      mLocationToWrite = savedInstanceState.getString(SI_LOCATION_TO_WRITE);
-    }
 
     final FragmentManager fm = getFragmentManager();
     final Fragment fragment = fm.findFragmentById(R.id.content);
@@ -57,9 +42,7 @@ public class QuranPreferenceActivity extends QuranActionBarActivity {
 
   @Override
   protected void onSaveInstanceState(Bundle outState) {
-    if (mLocationToWrite != null) {
-      outState.putString(SI_LOCATION_TO_WRITE, mLocationToWrite);
-    }
+
     super.onSaveInstanceState(outState);
   }
 
@@ -79,33 +62,4 @@ public class QuranPreferenceActivity extends QuranActionBarActivity {
     startActivity(intent);
   }
 
-  public void requestWriteExternalSdcardPermission(String newLocation) {
-    if (PermissionUtil.canRequestWriteExternalStoragePermission(this)) {
-      QuranSettings.getInstance(this).setSdcardPermissionsDialogPresented();
-
-      mLocationToWrite = newLocation;
-      ActivityCompat.requestPermissions(this,
-          new String[]{ Manifest.permission.WRITE_EXTERNAL_STORAGE },
-          REQUEST_WRITE_TO_SDCARD_PERMISSION);
-    } else {
-      // in the future, we should make this a direct link - perhaps using a Snackbar.
-      Toast.makeText(this, R.string.please_grant_permissions, Toast.LENGTH_SHORT).show();
-    }
-  }
-
-  @Override
-  public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
-      @NonNull int[] grantResults) {
-    if (requestCode == REQUEST_WRITE_TO_SDCARD_PERMISSION) {
-      if (grantResults.length == 1 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-        if (mLocationToWrite != null) {
-          Fragment fragment = getFragmentManager().findFragmentById(R.id.content);
-          if (fragment instanceof QuranSettingsFragment) {
-            ((QuranSettingsFragment) fragment).moveFiles(mLocationToWrite);
-          }
-        }
-      }
-      mLocationToWrite = null;
-    }
-  }
 }
