+++ app/src/main/res/xml/quran_advanced_preferences.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+      android:key="@string/prefs_path"
+      android:title="@string/prefs_category_advanced">
+  <PreferenceCategory
+      android:key="@string/prefs_advanced_category_key"
+      android:title="@string/prefs_category_advanced">
+
+    <com.quran.labs.androidquran.ui.preference.QuranPreference
+        android:key="@string/prefs_import"
+        android:title="@string/prefs_import_title"
+        android:summary="@string/prefs_import_summary"/>
+
+    <com.quran.labs.androidquran.ui.preference.QuranPreference
+        android:key="@string/prefs_export"
+        android:title="@string/prefs_export_title"
+        android:summary="@string/prefs_export_summary"/>
+
+    <com.quran.labs.androidquran.ui.preference.DataListPreference
+        android:key="@string/prefs_app_location"
+        android:summary="@string/prefs_app_location_summary"
+        android:title="@string/prefs_app_location_title"/>
+
+    <com.quran.labs.androidquran.ui.preference.QuranPreference
+        android:key="@string/prefs_send_logs"
+        android:title="@string/prefs_send_logs_title"
+        android:summary="@string/prefs_send_logs_summary"/>
+  </PreferenceCategory>
+</PreferenceScreen>
+
+
