+++ app/src/main/res/xml/quran_preferences.xml
@@ -146,33 +146,17 @@
         android:summary="@string/prefs_audio_manager_summary"/>
   </PreferenceCategory>
 
-  <PreferenceScreen
+  <PreferenceCategory
       android:key="@string/prefs_path"
       android:title="@string/prefs_category_advanced">
-  <PreferenceCategory
-      android:key="@string/prefs_advanced_category_key"
+<Preference
+      android:key="@string/prefs_path"
       android:title="@string/prefs_category_advanced">
-
-    <com.quran.labs.androidquran.ui.preference.QuranPreference
-        android:key="@string/prefs_import"
-        android:title="@string/prefs_import_title"
-        android:summary="@string/prefs_import_summary"/>
-
-    <com.quran.labs.androidquran.ui.preference.QuranPreference
-        android:key="@string/prefs_export"
-        android:title="@string/prefs_export_title"
-        android:summary="@string/prefs_export_summary"/>
-
-    <com.quran.labs.androidquran.ui.preference.DataListPreference
-        android:key="@string/prefs_app_location"
-        android:summary="@string/prefs_app_location_summary"
-        android:title="@string/prefs_app_location_title"/>
-
-    <com.quran.labs.androidquran.ui.preference.QuranPreference
-        android:key="@string/prefs_send_logs"
-        android:title="@string/prefs_send_logs_title"
-        android:summary="@string/prefs_send_logs_summary"/>
-  </PreferenceCategory>
-</PreferenceScreen>
+    <intent
+        android:targetPackage="com.quran.labs.androidquran"
+        android:targetClass="com.quran.labs.androidquran.AdvancedQuranPreferenceActivity"
+        android:action="com.example.prefs.PREFS_ONE" />
+  </Preference>
+</PreferenceCategory>
 </PreferenceScreen>
